Specify font-family for buttons without Helvetica to fix alignment
Using Helvetica for button fonts causes some weird vertical alignment
issues in Chrome OSX (FireFox and Safari are fine) for form elements such
as `<input>` and `<button>`, but not for other elements such as `<a>`.
Their alignment is especially noticeable if you place them next
to each other.

Removing Helvetica from the fonts list for these elements causes the
styling to fall back to using Arial, which fixes the issue. At the
button's font sizes, it's hard to tell the difference between the two
types anyway, and arguably even (barely) improves legibility.