Fix code style issues in SCSS source files
Respects code style as noted on http://primercss.io/guidelines/#css.
The bulk of this commit includes moving single property declarations to
a new line instead of having them on the same line as the opening and
closing brace.

One additional fix is to fix a faulty indentation in scss/_menu.scss.