Add utility class to truncate beginning of long text
Currently, we have the [`.css-truncate` 
class](https://primer.style/css/components/truncate), which shortens the 
end of long text with an ellipsis. Sometimes though, it makes sense to 
truncate the beginning of long text. For example, in the case of a file 
path, often the most important part is the file name, so we'd want to 
truncate the beginning of a long file path to make sure we display the 
name at the end of the path.

This change adds a `.truncate-left` class that shortens long text with 
an ellipsis by truncating the beginning of the text instead of the end.