Use tabular numbers in tables
As tables are frequently used for comparing columns of numeric data (and have the ability to be right-aligned) we should make sure that we make it as easy as possibly. Toggling tabular numbers on givens you the ability to scan columns of numbers effectively, assuming the font supports it. If it doesn’t then this doesn’t have any effect, but is no worse than the current behaviour.